Added Activity Logger

+++ actions/workspace.actions.ts
@@ -1,14 +1,17 @@
 "use server";
 
+import { activityLogger } from "@/lib/activity-logger";
+import { auth } from "@/lib/auth";
 import { executeAction } from "@/lib/executeAction";
 import prisma from "@/lib/prisma";
+import { ActivityInterface } from "@/types/types";
 import { revalidatePath } from "next/cache";
 
 export const addIssueAction = async (payload: any) => {
   const {
     title,
     description,
-    userId,
+    userId: assigneeId,
     priority,
     status,
     projectId,
@@ -19,16 +22,37 @@ export const addIssueAction = async (payload: any) => {
   return executeAction({
     successMessage: "Issue Added",
     actionFn: async () => {
+      const session = await auth();
+      const currentUser = session?.user?.id;
+
+      if (!currentUser) {
+        throw new Error(
+          "Unauthorized: You must be logged in to create an issue.",
+        );
+      }
+
       const createIssue = await prisma.issue.create({
         data: {
           title,
           description,
           priority,
           statusId: status,
-          assigneeId: userId,
+          assigneeId: assigneeId,
           projectId: projectId,
         },
       });
+      const loggerData: ActivityInterface = {
+        action: "CREATED",
+        entityTitle: title,
+        userId: currentUser,
+        workspaceId: workspaceId,
+        teamId: teamId,
+        projectId: projectId,
+        issueId: createIssue?.id,
+        beforeState: null,
+        afterState: null,
+      };
+      await activityLogger(loggerData);
       revalidatePath(`/${workspaceId}/team/${teamId}/project/${projectId}`);
       return createIssue;
     },
@@ -48,11 +72,20 @@ export const editIssueAction = async (payload: any) => {
     statusId,
   } = payload;
 
-  console.log(issueId);
-
   return executeAction({
-    successMessage: "Issue Added",
+    successMessage: "Issue Edited",
     actionFn: async () => {
+      const session = await auth();
+      const currentUser = session?.user?.id;
+
+      if (!currentUser) throw new Error("Unauthorized");
+
+      const oldIssue = await prisma.issue.findFirst({
+        where: { id: issueId },
+      });
+
+      if (!oldIssue) throw new Error("Issue not found.");
+
       const editIssue = await prisma.issue.update({
         where: { id: issueId },
         data: {
@@ -64,6 +97,7 @@ export const editIssueAction = async (payload: any) => {
           projectId,
         },
       });
+
       revalidatePath(`/${workspaceId}/team/${teamId}/project/${projectId}`);
       return editIssue;
     },
@@ -75,9 +109,33 @@ export const deleteIssue = async (payload: any) => {
   return executeAction({
     successMessage: "Issue Deleted",
     actionFn: async () => {
-      await prisma.issue.delete({
+      const session = await auth();
+      const currentUser = session?.user?.id;
+
+      if (!currentUser) {
+        throw new Error(
+          "Unauthorized: You must be logged in to create an issue.",
+        );
+      }
+
+      const issue = await prisma.issue.delete({
         where: { id: issueId },
       });
+
+      const loggerData: ActivityInterface = {
+        action: "DELETED",
+        entityTitle: issue.title,
+        userId: currentUser,
+        workspaceId: workspaceId,
+        teamId: teamId,
+        projectId: projectId,
+        issueId,
+        beforeState: null,
+        afterState: null,
+      };
+
+      await activityLogger(loggerData);
+      revalidatePath(`/${workspaceId}/dashboard`);
       revalidatePath(`/${workspaceId}/team/${teamId}/project/${projectId}`);
     },
   });
@@ -93,13 +151,48 @@ export const moveCardAction = async (payload: any) => {
   return executeAction({
     successMessage: "Card Moved",
     actionFn: async () => {
+      const session = await auth();
+      const currentUser = session?.user?.id;
+
+      if (!currentUser) {
+        throw new Error(
+          "Unauthorized: You must be logged in to create an issue.",
+        );
+      }
+
+      const fetchCard = await prisma.issue.findFirst({
+        where: { id: issueId },
+      });
+
+      const previousStatus = fetchCard?.statusId;
+
       const card = await prisma.issue.update({
         where: { id: issueId },
         data: {
           statusId: statusId,
         },
       });
+
       const projectId = card.projectId;
+
+      const loggerData: ActivityInterface = {
+        action: "STATUS_CHANGED",
+        entityTitle: card.title,
+        userId: currentUser,
+        workspaceId: workspaceId,
+        teamId: teamId,
+        projectId: projectId,
+        issueId: card?.id,
+        beforeState: {
+          previousIssueId: previousStatus,
+        },
+        afterState: {
+          newIssueId: card.statusId,
+        },
+      };
+
+      await activityLogger(loggerData);
+
       revalidatePath(`/${workspaceId}/team/${teamId}/project/${projectId}`);
       return card;
     },
